popfinder/backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
openai
python-dotenv
google-api-python-client
